pokemon/requirements.txt
djangorestframework-simplejwt==4.4.0
